implemented sorting images by color
-ANY two colors can be compared now
-BIG VCs now store values as floats instead of ints to accomodate for color
-fixed method headers, spelling/params
-removed unused resizer code
-added color-related methods in Util
-switched to using graphics2d for use with floats
-fixed a customization menu bug where all titles were "All Algorithms"

-branch almost complete

+++ src/main/java/main/vcs/VisualComponent.java
@@ -3,23 +3,23 @@ package main.vcs;
 public class VisualComponent
 {
 	//the value that is compared to other VCs
-	private int value;
+	private float value;
 	
 	/**
 	 * these are what the sorters hold in their arrays, the things being sorted
 	 * VC for short
 	 */
-	public VisualComponent(int value)
+	public VisualComponent(float value)
 	{
 		this.value = value;
 	}
 	
-	public final int getValue()
+	public final float getValue()
 	{
 		return value;
 	}
 	
-	public final void setValue(int value)
+	public final void setValue(float value)
 	{
 		this.value = value;
 	}
